Move even more REST files into src (#37041)
